updated unit tests, fixed dependency injection, bug fixes for the HBase DAO.

+++ JHouse/jhouse.core/src/main/java/jinvestor/jhouse/core/House.java
@@ -224,8 +224,8 @@ public class House {
 			this.acres = acres;
 			return this;
 		}
-		public HouseBuilder beds(byte beds) {
-			this.beds = beds;
+		public HouseBuilder beds(int beds) {
+			this.beds = (byte)beds;
 			return this;
 		}
 		public HouseBuilder baths(float baths) {
